projects/qa_mcp_bot/requirements.txt
huggingface_hub
duckduckgo-search
python-dotenv